Extract target step
