+++ src/components/theme-legacy/create-petition-form.js
@@ -7,9 +7,8 @@ import TargetCustom from './form/instructions/target-custom'
 import TargetNational from './form/instructions/target-national'
 import TargetState from './form/instructions/target-state'
 import Title from './form/instructions/title'
-import CustomTargetSelect from './form/target-select/custom'
-import NationalTargetSelect from './form/target-select/national'
-import StateTargetSelect from './form/target-select/state'
+
+import CreatePetitionTarget from '../../containers/create-petition-target'
 
 const instructionsByField = {
   title: <Title />,
@@ -23,22 +22,14 @@ const instructionsByField = {
 const CreatePetitionForm = ({
   selected,
   setSelected,
-  nationalOpen,
-  stateOpen,
-  customOpen,
   instructionStyle,
   setRef,
-  toggleOpen,
   errors,
   getValue,
   onChange,
   onPreview
 }) => {
   const instructions = instructionsByField[selected]
-
-  const national = !nationalOpen ? '' : <NationalTargetSelect />
-  const state = !stateOpen ? '' : <StateTargetSelect />
-  const custom = !customOpen ? '' : <CustomTargetSelect />
 
   return (
     <div className='container'>
@@ -88,73 +79,11 @@ const CreatePetitionForm = ({
                 />
               </div>
             </fieldset>
-            <div id='target_wrapper' className='' title='Choosing a Target'>
-              <fieldset id='target' className=''>
-                <span className='circle-number'>2</span>
-                <span className='lanky-header moveon-dark-blue'>
-                  Who&#39;s the target of your petition?
-                </span>
-                <div
-                  className='checkbox wrapper'
-                  id='national_group_wrapper'
-                  title='Targeting the White House or Congress'
-                  onClick={setSelected('target-national')}
-                >
-                  <label htmlFor='national_group' id='national_group_label'>
-                    <input
-                      name='checkbox_national_group'
-                      id='national_group'
-                      type='checkbox'
-                      className='reveal_more_options'
-                      onClick={toggleOpen('nationalOpen')}
-                      ref={setRef('nationalInput')}
-                    />{' '}
-                    The White House or Congress
-                  </label>
-                  {national}
-                </div>
-                <div
-                  className='checkbox wrapper'
-                  id='state_group_wrapper'
-                  title='Targeting Your Governor or State Legislature'
-                  onClick={setSelected('target-state')}
-                >
-                  <label htmlFor='state_group' id='state_group_label'>
-                    <input
-                      name='checkbox_state_group'
-                      id='state_group'
-                      type='checkbox'
-                      className='reveal_more_options'
-                      onClick={toggleOpen('stateOpen')}
-                      ref={setRef('stateInput')}
-                    />{' '}
-                    Your governor or state legislature
-                  </label>
-                  {state}
-                </div>
-                <div
-                  className='checkbox wrapper'
-                  id='checkbox_custom_group_wrapper'
-                  onClick={setSelected('target-custom')}
-                >
-                  <label
-                    htmlFor='custom_group'
-                    id='checkbox_custom_group_label'
-                  >
-                    <input
-                      name='checkbox_custom_group'
-                      id='custom_group'
-                      type='checkbox'
-                      className='reveal_more_options'
-                      onClick={toggleOpen('customOpen')}
-                      ref={setRef('customInput')}
-                    />{' '}
-                    Someone else (like a local official or corporate CEO)
-                  </label>
-                  {custom}
-                </div>
-              </fieldset>
-            </div>
+            <CreatePetitionTarget
+              setSelected={setSelected}
+              setRef={setRef}
+              onChange={onChange}
+            />
             <fieldset id='statement'>
               <span className='circle-number'>3</span>
               <span className='lanky-header moveon-dark-blue'>
@@ -200,12 +129,8 @@ const CreatePetitionForm = ({
 CreatePetitionForm.propTypes = {
   selected: PropTypes.string,
   setSelected: PropTypes.func,
-  nationalOpen: PropTypes.bool,
-  stateOpen: PropTypes.bool,
-  customOpen: PropTypes.bool,
   instructionStyle: PropTypes.object,
   setRef: PropTypes.func,
-  toggleOpen: PropTypes.func,
   errors: PropTypes.array.isRequired,
   getValue: PropTypes.func,
   onChange: PropTypes.func,
